moved create_one_to_many_link function
