+++ vantage6-algorithm-store/vantage6/algorithm/store/resource/schema/output_schema.py
@@ -1,47 +1,19 @@
 from marshmallow import fields
-from flask import url_for
 
-from vantage6.backend.common.resource.output_schema import BaseHATEOASModelSchema
+from vantage6.backend.common.resource.output_schema import (
+    BaseHATEOASModelSchema,
+    create_one_to_many_link
+)
 from vantage6.algorithm.store.model.algorithm import Algorithm
 from vantage6.algorithm.store.model.argument import Argument
 from vantage6.algorithm.store.model.database import Database
 from vantage6.algorithm.store.model.function import Function
-from vantage6.algorithm.store.model import Base
 from vantage6.algorithm.store.model.role import Role
 from vantage6.algorithm.store.model.rule import Rule
 from vantage6.algorithm.store.model.ui_visualization import UIVisualization
 from vantage6.algorithm.store.model.user import User
 from vantage6.algorithm.store.model.review import Review
 from vantage6.algorithm.store.model.vantage6_server import Vantage6Server
-
-
-# TODO: Remove duplicated code
-def create_one_to_many_link(obj: Base, link_to: str, link_from: str) -> str:
-    """
-    Create an API link to get objects related to a given object.
-
-    Parameters
-    ----------
-    obj : Base
-        Object to which the link is created
-    link_to : str
-        Name of the resource to which the link is created
-    link_from : str
-        Name of the resource from which the link is created
-
-    Returns
-    -------
-    str
-        API link
-
-    Examples
-    --------
-    >>> create_one_to_many_link(obj, "node", "organization_id")
-    "/api/node?organization_id=<obj.id>"
-    """
-    endpoint = link_to + "_without_id"
-    values = {link_from: obj.id}
-    return url_for(endpoint, **values)
 
 
 class HATEOASModelSchema(BaseHATEOASModelSchema):
