+++ vantage6-backend-common/vantage6/backend/common/resource/output_schema.py
@@ -3,11 +3,40 @@ from marshmallow_sqlalchemy import SQLAlchemyAutoSchema
 from sqlalchemy.orm.decl_api import DeclarativeMeta
 from flask import url_for
 
+from vantage6.backend.common.base import Base
 from vantage6.common import logger_name
 from vantage6.backend.common.resource.pagination import Pagination
 
 
 log = logging.getLogger(logger_name(__name__))
+
+
+def create_one_to_many_link(obj: Base, link_to: str, link_from: str) -> str:
+    """
+    Create an API link to get objects related to a given object.
+
+    Parameters
+    ----------
+    obj : Base
+        Object to which the link is created
+    link_to : str
+        Name of the resource to which the link is created
+    link_from : str
+        Name of the resource from which the link is created
+
+    Returns
+    -------
+    str
+        API link
+
+    Examples
+    --------
+    >>> create_one_to_many_link(obj, "node", "organization_id")
+    "/api/node?organization_id=<obj.id>"
+    """
+    endpoint = link_to + "_without_id"
+    values = {link_from: obj.id}
+    return url_for(endpoint, **values)
 
 
 class BaseHATEOASModelSchema(SQLAlchemyAutoSchema):
